Split the subscriber and the base for Spatial... not sure this is the right idiom
