+++ src/components/spatial/Spatial.js
@@ -4,53 +4,13 @@ import DeckGL, {ScatterplotLayer, PolygonLayer, COORDINATE_SYSTEM, OrthographicV
 import {Matrix4} from 'math.gl';
 import {BitmapLayer} from '@deck.gl/experimental-layers';
 import PropTypes from 'prop-types';
-import PubSub from 'pubsub-js';
-import { IMAGE_ADD, MOLECULES_ADD } from '../events';
 
-export const INITIAL_VIEW_STATE = {
-  zoom: 2, // TODO: zoom=3 or above does not work?
-  maxZoom: 40, // TODO: derive from user data
-  pitch: 0,
-  bearing: 0,
-  offset: [10000, 10000] // Required: https://github.com/uber/deck.gl/issues/2580
-  // TODO: derive from user data
-};
-
-export class SpatialSubscriber extends React.Component {
-  constructor(props) {
-    super(props);
-    this.state = {baseImgUrl: undefined};
-  }
-
-  componentWillMount() {
-    this.imageToken = PubSub.subscribe(IMAGE_ADD, this.imageAddSubscriber.bind(this));
-    this.moleculesToken = PubSub.subscribe(MOLECULES_ADD, this.moleculesAddSubscriber.bind(this));
-  }
-
-  componentWillUnmount() {
-    PubSub.unsubscribe(this.imageToken);
-    PubSub.unsubscribe(this.moleculesToken);
-  }
-
-  imageAddSubscriber(msg, baseImg) {
-    this.setState({baseImg: baseImg});
-  }
-
-  moleculesAddSubscriber(msg, molecules) {
-    this.setState({molecules: molecules});
-  }
-
-  render() {
-    return (
-      <Spatial baseImg={this.state.baseImg} molecules={this.state.molecules}/>
-    );
-  }
-}
 
 function renderLayers(props) {
   const {
     baseImg = undefined,
-    molecules = undefined
+    molecules = undefined,
+    cells = undefined
   } = props;
 
   const polygon_data = [ { contour: [[-20, -20], [-65, -10], [-80, 0], [-70, 40]] } ];
@@ -144,8 +104,17 @@ function renderLayers(props) {
   return layers;
 }
 
-export function Spatial(props) {
+export default function Spatial(props) {
   const {viewState, controller = true} = props;
+
+  const INITIAL_VIEW_STATE = {
+    zoom: 2, // TODO: zoom=3 or above does not work?
+    maxZoom: 40, // TODO: derive from user data
+    pitch: 0,
+    bearing: 0,
+    offset: [10000, 10000] // Required: https://github.com/uber/deck.gl/issues/2580
+    // TODO: derive from user data
+  };
 
   return (
     <DeckGL
