+++ src/components/spatial/index.js
@@ -1,0 +1,1 @@
+export { SpatialSubscriber } from './SpatialSubscriber';
